requirements.txt
python-dotenv
openai
gradio
git+https://github.com/openai/whisper.git 
